Ajout des permissions 1er commit

+++ module/Ent/src/Ent/Controller/ModuleController.php
@@ -35,6 +35,10 @@ class ModuleController extends AbstractActionController
 
     public function listAction()
     {
+        if (!$this->isGranted('list_module')) {
+            throw new \ZfcRbac\Exception\UnauthorizedException('You are not allowed !');
+        }
+        
         $listModules = $this->moduleService->getAll();
 
         return new ViewModel(array(
@@ -44,6 +48,10 @@ class ModuleController extends AbstractActionController
 
     public function addAction()
     {
+        if (!$this->isGranted('add_module')) {
+            throw new \ZfcRbac\Exception\UnauthorizedException('You are not allowed !');
+        }
+        
         $form = $this->moduleForm;
 
         if ($this->request->isPost()) {
@@ -52,7 +60,7 @@ class ModuleController extends AbstractActionController
             if ($module) {
                 $this->flashMessenger()->addSuccessMessage('Le module a bien été ajouté.');
 
-                return $this->redirect()->toRoute('module');
+                return $this->redirect()->toRoute('zfcadmin/module');
             }
         }
 
@@ -63,6 +71,10 @@ class ModuleController extends AbstractActionController
 
     public function showAction()
     {
+        if (!$this->isGranted('show_module')) {
+            throw new \ZfcRbac\Exception\UnauthorizedException('You are not allowed !');
+        }
+        
         $id = $this->params('id');
 
         $module = $this->moduleService->getById($id);
@@ -78,6 +90,10 @@ class ModuleController extends AbstractActionController
 
     public function updateAction()
     {
+        if (!$this->isGranted('update_module')) {
+            throw new \ZfcRbac\Exception\UnauthorizedException('You are not allowed !');
+        }
+        
         $id = $this->params('id');
         $form = $this->moduleForm;
         $module = $this->moduleService->getById($id, $form);
@@ -88,7 +104,7 @@ class ModuleController extends AbstractActionController
             if ($module) {
                 $this->flashMessenger()->addSuccessMessage('Le module a bien été modifié.');
 
-                return $this->redirect()->toRoute('module');
+                return $this->redirect()->toRoute('zfcadmin/module');
             }
         }
 
@@ -99,13 +115,17 @@ class ModuleController extends AbstractActionController
 
     public function deleteAction()
     {
+        if (!$this->isGranted('delete_module')) {
+            throw new \ZfcRbac\Exception\UnauthorizedException('You are not allowed !');
+        }
+        
         $id = $this->params('id');
 
         $this->moduleService->delete($id);
 
         $this->flashMessenger()->addSuccessMessage('Le module a bien été supprimé.');
 
-        return $this->redirect()->toRoute('module');
+        return $this->redirect()->toRoute('zfcadmin/module');
     }
 
 }
